Challenge Days Between Days
Here is the challenge for days between days.  I did this different then he did. so if I need to know how he did it or if his way is best practice I'll need to go back and watch his solution which is the last challenge in lesson 4 on Developer University web page